remove DatabaseProvider and move AppDatabase
