--- app/src/main/java/com/jin/jjinweather/feature/database/data/DatabaseProvider.kt
@@ -1,19 +1,0 @@
-package com.jin.jjinweather.feature.database.data
-
-import android.content.Context
-import androidx.room.Room
-
-object DatabaseProvider {
-    @Volatile
-    private var INSTANCE: AppDatabase? = null
-
-    fun getDatabase(context: Context): AppDatabase {
-        return INSTANCE ?: synchronized(this) {
-            INSTANCE ?: Room.databaseBuilder(
-                context.applicationContext,
-                AppDatabase::class.java,
-                "weather_db"
-            ).build().also { INSTANCE = it }
-        }
-    }
-}
